Eigen: Added non linear functor
git-svn-id: svn://ultimatepp.org/upp/trunk@6022 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/plugin/Eigen/Eigen/src/QR/ColPivHouseholderQR.h
@@ -458,7 +458,7 @@ struct solve_retval<ColPivHouseholderQR<_MatrixType>, Rhs>
   {
     eigen_assert(rhs().rows() == dec().rows());
 
-    const int cols = dec().cols(),
+    const ptrdiff_t cols = dec().cols(),
     nonzero_pivots = dec().nonzeroPivots();
 
     if(nonzero_pivots == 0)
